requirements.txt
crewai==0.11.0
langchain-openai==0.0.2.post1
langchain-community==0.0.13
langchain==0.1.0
python-dotenv==1.0.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.1
requests==2.31.0
beautifulsoup4==4.12.2
openai>=1.7.1,<2.0.0
duckduckgo-search>=4.1.1
wikipedia==1.4.0